+++ benchopt/runner.py
@@ -92,15 +92,6 @@ def run_one_to_cvg(benchmark, objective, solver, meta, stopping_criterion,
     with exception_handler(output, pdb=pdb) as ctx:
 
         if solver._solver_strategy == "callback":
-            output.progress('empty run for compilation')
-            run_once_cb = _Callback(
-                lambda x: {'objective_value': 1},
-                {},
-                stopping_criterion.get_runner_instance(
-                    solver=solver, max_runs=1
-                )
-            )
-            solver.run(run_once_cb)
 
             # If stopping strategy is 'callback', only call once to get the
             # results up to convergence.
@@ -180,9 +171,8 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
         output.skip(reason, objective=True)
         return []
 
-    skip, reason = solver._set_objective(objective)
+    skip = solver._set_objective(objective, output=output)
     if skip:
-        output.skip(reason)
         return []
 
     states = []