+++ doc/whats_new.rst
@@ -16,6 +16,13 @@ CLI
 API
 ---
 
+- Add :class:`~benchopt.stopping_criterion.SingleRunCriterion` to run a solver
+  only once. This can be used for benchmarking methods where we are interested
+  in objective value at convergence. By `Thomas Moreau`_ (:gh:`511`)
+- Add :func:`~benchopt.BaseSolver.run_once` helper to easily warmup solvers
+  with callback. By `Thomas Moreau`_ (:gh:`511`)
+
+
 FIX
 ---